added connections
