+++ Buildoapi/views/__init__.py
@@ -4,4 +4,4 @@ from .workorders import WorkOrderViewSet
 from .reviews import ReviewViewSet
 from .contractors import ContractorViewSet
 from .job_requests import JobRequestViewSet
-from .connections import ConnectionRequestViewSet
+from .connection_request import ConnectionRequestViewSet
